Create avoiding_obstacle_bug0.ino
The main idea of obstacle avoidance is bug 0 algorithm. The sensor applied is Ultrasonic Sensor - HC-SR04. When the sensor detects obstacle in the front within one and a half meter, the robot will turn left. If there is an obstacle on the right, the robot will go forward to follow the obstacle. When the robot is able to turn directly to the target, it will determine the turning direction according to the difference between the current GPS and the GPS location of the target location. The detail is shown in the section of target navigation. In the code, the target angle is preset.