create target_navigation.ino
Target navigation is largely dependent on the bearing angle between target GPS position and current GPS position. After knowing the bearing angle, the robot will compare its own orientation angle (the angle between the heading of the robot and the north) with the target angle and turn according to the bearing angle. 
Current GPS position is updated every 0.2 seconds. The target angle (bearing angle) is calculated by the formula as followed.
	X=cos(target_latitude)*sin(target_longitude-current_longitude)
	Y=cos(current_latitude)*sin(target_latitude)-sin(current_latitude)*cos(target_latitude)*cos(target_longitude-current_longitude)
	Target angle= atan2(X,Y)*180/PI
Current orientation angle is from Adafruit BNO055 Absolute Orientation Sensor. It is updated in a frequency of 15 Hz. The orientation angle ranges from 0 to 360 degrees. 
Comparing the orientation angle and the target angle (target angle – current angle):
-	-5 ~ 5 degrees: go straight forward
-	5~180 degrees: turning right
-	180 ~ 355 degrees: turning left
-	-180~-5 degrees: turning left
-	-355 ~ -180: turning right
Once the robot reached the target position, it will stop.